opt account & proxy

+++ admins/services/accountservice.go
@@ -1,13 +1,77 @@
 package services
 
 import (
+	"context"
+	"sync"
 	"time"
 
+	"github.com/juggleim/commons/caches"
+	"github.com/juggleim/commons/ctxs"
 	"github.com/juggleim/commons/errs"
 	utils "github.com/juggleim/commons/tools"
 	apimodels "github.com/juggleim/jugglechat-server/admins/apis/models"
 	"github.com/juggleim/jugglechat-server/storages/dbs"
 )
+
+type RoleType int
+type AccountState int
+
+const (
+	RoleType_SuperAdmin  RoleType = 0
+	RoleType_NormalAdmin RoleType = 1
+
+	AccountState_Normal AccountState = 0
+)
+
+type AccountInfo struct {
+	Account  string
+	RoleType RoleType
+	State    AccountState
+}
+
+var accountInfoCache *caches.LruCache
+var accountLock *sync.RWMutex
+
+func init() {
+	accountLock = &sync.RWMutex{}
+	accountInfoCache = caches.NewLruCacheWithAddReadTimeout("account_cache", 100, nil, 10*time.Minute, 10*time.Minute)
+}
+
+var notExistAccount = &AccountInfo{}
+
+func GetAccountInfo(account string) (*AccountInfo, bool) {
+	if val, exist := accountInfoCache.Get(account); exist {
+		info := val.(*AccountInfo)
+		if info == notExistAccount {
+			return nil, false
+		}
+		return info, true
+	} else {
+		accountLock.Lock()
+		defer accountLock.Unlock()
+		if val, exist := accountInfoCache.Get(account); exist {
+			info := val.(*AccountInfo)
+			if info == notExistAccount {
+				return nil, false
+			}
+			return info, true
+		} else {
+			dao := dbs.AccountDao{}
+			acc, err := dao.FindByAccount(account)
+			if err != nil || acc == nil {
+				accountInfoCache.Add(account, notExistAccount)
+				return nil, false
+			}
+			info := &AccountInfo{
+				Account:  account,
+				State:    AccountState(acc.State),
+				RoleType: RoleType(acc.RoleType),
+			}
+			accountInfoCache.Add(account, info)
+			return info, true
+		}
+	}
+}
 
 func CheckLogin(account, password string) (errs.AdminErrorCode, *apimodels.Account) {
 	dao := dbs.AccountDao{}
@@ -69,7 +133,15 @@ func UpdPassword(account, password, newPassword string) errs.AdminErrorCode {
 	return errs.AdminErrorCode_Success
 }
 
-func AddAccount(parentAccount, account, password string, roleType int) errs.AdminErrorCode {
+func AddAccount(ctx context.Context, account, password string, roleType int) errs.AdminErrorCode {
+	parentAccount := ctxs.GetAccountFromCtx(ctx)
+	curAccount, exist := GetAccountInfo(parentAccount)
+	if !exist || curAccount == nil {
+		return errs.AdminErrorCode_AccountNotExist
+	}
+	if curAccount.RoleType != RoleType_SuperAdmin {
+		return errs.AdminErrorCode_NotPermission
+	}
 	dao := dbs.AccountDao{}
 	password = utils.SHA1(password)
 	err := dao.Create(dbs.AccountDao{
@@ -87,18 +159,72 @@ func AddAccount(parentAccount, account, password string, roleType int) errs.Admi
 	return errs.AdminErrorCode_Success
 }
 
-func DisableAccounts(accounts []string, isDisable int) errs.AdminErrorCode {
+func DisableAccounts(ctx context.Context, accounts []string, isDisable int) errs.AdminErrorCode {
+	curAccount, exist := GetAccountInfo(ctxs.GetAccountFromCtx(ctx))
+	if !exist || curAccount == nil {
+		return errs.AdminErrorCode_AccountNotExist
+	}
+	if curAccount.RoleType != RoleType_SuperAdmin {
+		return errs.AdminErrorCode_NotPermission
+	}
 	dao := dbs.AccountDao{}
 	dao.UpdateState(accounts, isDisable)
 	return errs.AdminErrorCode_Success
 }
-func DeleteAccounts(accounts []string) errs.AdminErrorCode {
+
+func DeleteAccounts(ctx context.Context, accounts []string) errs.AdminErrorCode {
+	curAccount, exist := GetAccountInfo(ctxs.GetAccountFromCtx(ctx))
+	if !exist || curAccount == nil {
+		return errs.AdminErrorCode_AccountNotExist
+	}
+	if curAccount.RoleType != RoleType_SuperAdmin {
+		return errs.AdminErrorCode_NotPermission
+	}
 	dao := dbs.AccountDao{}
 	dao.DeleteAccounts(accounts)
 	return errs.AdminErrorCode_Success
 }
 
-func QryAccounts(limit int64, offset string) *apimodels.Accounts {
+func BindApps(ctx context.Context, account string, appkeys []string) errs.AdminErrorCode {
+	curAccount, exist := GetAccountInfo(ctxs.GetAccountFromCtx(ctx))
+	if !exist || curAccount == nil {
+		return errs.AdminErrorCode_AccountNotExist
+	}
+	if curAccount.RoleType != RoleType_SuperAdmin {
+		return errs.AdminErrorCode_NotPermission
+	}
+	dao := dbs.AccountAppRelDao{}
+	for _, appkey := range appkeys {
+		dao.Create(dbs.AccountAppRelDao{
+			AppKey:  appkey,
+			Account: account,
+		})
+	}
+	return errs.AdminErrorCode_Success
+}
+
+func UnBindApps(ctx context.Context, account string, appkeys []string) errs.AdminErrorCode {
+	curAccount, exist := GetAccountInfo(ctxs.GetAccountFromCtx(ctx))
+	if !exist || curAccount == nil {
+		return errs.AdminErrorCode_AccountNotExist
+	}
+	if curAccount.RoleType != RoleType_SuperAdmin {
+		return errs.AdminErrorCode_NotPermission
+	}
+	dao := dbs.AccountAppRelDao{}
+	dao.BatchDelete(account, appkeys)
+	return errs.AdminErrorCode_Success
+}
+
+func QryAccounts(ctx context.Context, limit int64, offset string) (errs.AdminErrorCode, *apimodels.Accounts) {
+	curAccount, exist := GetAccountInfo(ctxs.GetAccountFromCtx(ctx))
+	if !exist || curAccount == nil {
+		return errs.AdminErrorCode_AccountNotExist, nil
+	}
+	if curAccount.RoleType != RoleType_SuperAdmin {
+		return errs.AdminErrorCode_NotPermission, nil
+	}
+
 	accounts := &apimodels.Accounts{
 		Items:   []*apimodels.Account{},
 		HasMore: false,
@@ -135,5 +261,5 @@ func QryAccounts(limit int64, offset string) *apimodels.Accounts {
 			accounts.Offset = offset
 		}
 	}
-	return accounts
+	return errs.AdminErrorCode_Success, accounts
 }
